requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
attrs==24.3.0
black==24.10.0
cachetools==5.5.0
certifi==2024.12.14
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
deprecation==2.1.0
distlib==0.3.9
distro==1.9.0
Django==5.1.4
django-cors-headers==4.6.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
filelock==3.16.1
flake8==7.1.1
frozenlist==1.5.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
gotrue==2.11.1
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
hyperframe==6.0.1
identify==2.6.4
idna==3.10
inflection==0.5.1
isort==5.13.2
jiter==0.8.2
mccabe==0.7.0
multidict==6.1.0
mypy-extensions==1.0.0
nodeenv==1.9.1
openai==1.58.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
postgrest==0.19.1
pre_commit==4.0.1
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pydantic==2.10.4
pydantic_core==2.27.2
pyflakes==3.2.0
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
realtime==2.1.0
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
storage3==0.11.0
StrEnum==0.4.15
supabase==2.11.0
supafunc==0.9.0
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
virtualenv==20.28.0
websockets==13.1
yarl==1.18.3
